requirements.txt
langchain==0.0.1  
PyMuPDF==1.23.3  
faiss-cpu==1.7.3
transformers==4.43.3
langchain_groq==0.1.9
python-dotenv==1.0.1
gTTS==2.2.4
streamlit==1.37.0
langchain-community==0.2.10
langchain-text-splitters==0.2.2
unstructured==0.15.0
unstructured[pdf]==0.15.0
sentence-transformers==3.0.1
nltk==3.8.1
